connect up PlayQueueList actions and remove old NowPlayingPage
